Unzip Windows ExifTool executable using zip::unzip, not utils::unzip
The move to `zip::unzip()` is an attempt to avoid problems installing
the the executable using `install_exiftool()`. User @Nova-Scotia,
teaching a class in which she used the package, reported that for many
students, `install_exiftool()` did not "just work". She also thought
that the unzipping step might have been the location of some of their
problems. The zip provides a standalone executable that's consistent
among operating systems (and, presumably, on different Windows
machines). Since it works fine on all the Windows computers I've
tested it on, I hope that this will make `install_exiftool()` behave
more consistently on others' machines.