Option to have exiftool (in intermediate step) output results in csv
As documented at https://exiftool.org/exiftool_pod.html, "the JSON
output is UTF-8 regardless of any -L or -charset option setting". This
can be problematic when metadata are recorded in some character sets.

If, instead of setting the "-j" flag, one sets the "-T" and "-csv"
flags, exiftool will output metadata in a tabular csv, and will
respect any -charset flags. We can then read in the data to R from the
csv, rather than from a json struction. To offer this alternative
intermediate, I have added the option of setting `pipeline="csv"`.

The csv pipeline, though, has the significant downside. CSVs drop any
explicit information about the types of their columns, leaving the R
to guess as the types as it imports them from the csv. This can have
bad side effects, such as a version string formatted as "0100" being
read in as the number 100, and (with read.csv) a column filled with
the character "T" being read in as a column of Boolean TRUE values. To
avoid this, I've added an option `pipeline="both"` that first reads
the data and outputs them as a JSON structure (to get the column
classes) and then outputs the data as a csv, which it reads in using
the just extracted classes in the colClasses= argument. This option
gives you (when you need them) the best of both worlds, at the expense
of having to call exiftool twice instead of once.