Fix problem on Linux with exif_call()
Issue #4 describes the failure of `exif_call()` on a Linux
machine. That problem results because the value of `exiftoolpath`
returned by get_exiftool_command() is of this form:

```r
"'/path/to/perl' '/path/to/exiftool'"
```

Formerly (prior to commit a24e28f012dd8 on 2020-10-22), `exif_call()`
used `system(command, ...)` to launch its external call to
ExifTool. That handled an `exiftoolpath` like the one above just fine.

In commit a24e28f012dd8, `exif_call()` shifted from `system()` to
`system2(exiftoolpath, args=args, stdout=stdout)`. The switch to
`system2()` was needed because the then newly introduced
`pipeline="csv"` option needs to be able to write stdout to a file,
and `system2()` supports that, while `system()` does not.
Unfortunately, `system2()` fails with values of `exiftoolpath` like
the one above, since it only accepts a single token in the string
passed as its first argument.

This commit fixes that issue by inspecting, in
`configure_exiftoolr()`, any value of `command` that succeeded in
launching ExifTool and, if it contains more than one token, splitting
it into a character vector before passing it on. If passed a value of
`exiftoolpath` of length greater than one, `exif_call()` now passes
its first element to `system2()` as its `command=` argument, and
prepends the remaining elements to the character vector passed to its
`args=` argument.