Use system2() throughout to handle spaces in paths to executables
In issue #5, @LafontRapnouilTristan reported failure of
`configure_exiftoolr()` on a Windows 10 machine even though
`install_exiftool()` had succeeded in installing the exiftool
executable in the package directory tree. Running
`system2("/path/to/exiftool(-k).exe", "-ver")` succeeded, leading me
to poke around and find that `configure_exiftoolr()` (which used
`system()`) would fail when there are any spaces in the path to the
exiftool executable.

This commit fixes that problem, at least on Windows. Will also test it
on a Linux box before pushing the fix along to CRAN.