Add a (MacOS-related) exiftool invocation to configure_exiftoolr
Based on Issue #16, it looks like on MacOS (at least), the exiftool
script needs to be run without a prepended call to perl. This commit
tests that theory, by adding a direct call to the exiftool script that
gets installed by `exiftoolr:::download_exiftool()`.

Based on a glance at the exiftool script (which begins with a
`#!/usr/bin/perl -w`), it seems possible that a prepended `perl` isn't
now needed on *NIX OS's. If that's the case, I may at some point
remove that version of the incantation from the set I currently test,
but as far as I can tell, it does no harm to leave it for now.