Add config_file option to exif_call()
Addresses #19, in which @burnsal pointed out that `exif_call()`
errored out when it passed in `-config CONFIGFILE` via the `args=`
argument.

The problem is that (as documented
[here](https://exiftool.org/exiftool_pod.html)):

> Arguments in ARGFILE behave exactly the same as if they were entered
> at the location of the -@ option on the command line, with the
> exception that the -config and -common_args options may not be used in
> an ARGFILE.

Because `exif_call()` writes the options in `args=` to a text file
which it then passes to `Exiftool` via its `-@` option, the config
file can't be specified via `args=`. It may instead be passed in via
the newly added `config_file=` argument.