Handle case in which ExifTool finds no files to read
As reported in Issue #20, `exif_read()` produces hard-to-decipher
errors in cases where the underlying call to ExifTool finds no
files (as, for example, when they are all filtered out by the user-set
`--ext` value). This commit adds code to handle that situation by:
(1) catching the error or warning message that ExifTool produces and
passing it on as an R warning; and (2) returning a value of `NULL`.